+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Variable type `time` has been updated to `time_of_day` for clarity
 - Improved consistency of internal data model, renamed Sheet and Grid `response`
   to `value`
+- Refactored sheet model and coverage methods
 
 ## 0.48.1 (February 13, 2017)